promote Tz::JmSynthe::isEnableMaterial; auto-correct header return types
Round 8 reported 0 matched, but one of its parked rows was byte-exact:
0x002a74c0 reached 100.0 % on attempt 8 and was parked as "attempt budget
exhausted". The driver had scored it before the registry gained the stub callee
rows added earlier today, so the mini-link could not resolve it at the time.

It failed to promote for the same reason YS::ACT::callback did: jm_synthe.hpp
declared isEnableMaterial as returning u64, the definition returns u8. Return
types are not mangled, so the two merely have to agree -- but which one is
right is settled by the compiled bytes, and u64 makes gcc emit a widening
sequence that drops the function to 97.78 %. The header was wrong; corrected,
with the evidence in a comment beside it.

fix_return_type_mismatch() now does this automatically: on a "prototype ... does
not match any in class" failure where the argument lists are identical and only
the return type differs, it rewrites the header declaration to match the
already-proven-exact definition and verifies once more. A real signature
difference still fails, and every touched header is tracked for rollback.

That is the second header this has caught today (act.hpp declared a bool
return that cost 4.35 %), so it is a class of repo error worth automating.

+++ src/tozawa/jm_synthe.hpp
@@ -6,7 +6,9 @@ namespace Tz {
 
 class JmSynthe {
 public:
-    u64 isEnableMaterial(void);
+    // u8, not u64: the original returns the byte it loaded, unwidened.
+    // Declaring u64 makes gcc emit a widening sequence (97.78 %).
+    u8 isEnableMaterial(void);
     bool isEnableRecipe(void);
     void CreateList(void);
     void Init(void);
